relocating TaskRow and its container
