+++ src/features/task/components/row/TaskRowContainer.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { TaskProps } from '@/shared/types/api.type';
-import { TaskRow } from '../components/TaskRow';
+import { TaskRow } from './TaskRow';
 import { showError } from '@/shared/services/notification/notificationService';
 import { useQueryClient } from '@tanstack/react-query';
